Output all types to the types folder
