+++ .gitignore
@@ -1,13 +1,13 @@
-node_modules/
 .DS_Store
+node_modules/
 /build/
-/web-components/
-/storyHelpers/
+/icons-raw/
+/icons-sf
+/icons-skia
 /react/
 /shared/
-/svelte/
-/tokens/
-/icons-raw/
 /storybook-static/
-/icons-skia
-/icons-sf
+/storyHelpers/
+/tokens/
+/types/
+/web-components/
